+++ changelog.md
@@ -118,3 +118,4 @@
 * move ALL image urls from `upload.wikimedia.org/wikipedia/commons` to `wikipedia.org/wiki/Special:Redirect/file/` via #86
 * image captions are now Sentence objects
 * rename citation → reference internally, and in json output
+* remove references inside section titles